requirements.txt
Django>=3.0,<4.0

djangorestframework~=3.14.0
pip~=22.2.2
attrs~=21.2.0
distro~=1.7.0
wheel~=0.37.1
cryptography~=3.4.8
Pillow~=9.0.1
keyring~=23.5.0
pytz~=2022.1
asgiref~=3.5.2
sqlparse~=0.4.2
colorama~=0.4.4
PyYAML~=5.4.1
requests~=2.25.1
setuptools~=59.6.0
pyparsing~=2.4.7
zipp~=1.0.0
future~=0.18.2